requirements/requirements.txt
Django
django-rest-framework
selenium
beautifulsoup4
lxml
